Completed CRUD operations of user by Admin. Added search user feature in admin panel. Separated user login and admin login into two separate routes.

+++ server/controller/controller.js
@@ -1,7 +1,7 @@
 const Userdb = require('../model/model');
 const bcrypt = require('bcrypt');
 
-//create user
+//create user 
 exports.create = async (req, res) => {
     if(!req.body){
         console.log("No body submitted");
@@ -14,23 +14,27 @@ exports.create = async (req, res) => {
         gender: req.body.genderRadio,
         phone: req.body.phone,
         email: req.body.email,
+        updatedAt: Date.now(),
         password: hashedPassword
     })
     console.log(user);
     user.save(user)
         .then(data => {
             console.log("Added user data: " + data)
-            res.redirect('/');
-        })
+            if(req.session.isAdmin)
+                res.redirect('/admin/users');
+            else
+                res.redirect('/');
+            })
         .catch(err => {
             res.status(500).render('error', {
-                message: err.message || "Unable to add to database"
+                message: err.message || "Unable to add user to database"
             });
         });
     }
 };
 
-//find one user by ID
+//find one user by ID 
 exports.findOne = (id) => {
     return new Promise((resolve, reject) => {
         Userdb.findById(id).lean()
@@ -52,27 +56,7 @@ exports.findOne = (id) => {
 
 //find and retrieve user(s)
 exports.find = (req, res) => {
-    // if(req.query.id){
-    //     const id = req.query.id;
-    //     Userdb.findById(id).lean()
-    //         .then(data => {
-    //             if(!data){
-    //                 res.status(404).render('error',{
-    //                     message: err.message || "Unable to find user"
-    //                 });
-    //             }
-    //             else{
-    //                 console.log(data);
-    //                 res.redirect('/');
-    //             }
-    //         })
-    //         .catch(err =>{
-    //             res.status(500).render('error',{
-    //                 message: err.message || "Error finding   user in Database"
-    //             });
-    //         }); 
-    // }
-    // else{
+    
     Userdb.find({ isAdmin: false }).sort({ name: 1 }).lean()
         .then(user => {
             console.log(user);
@@ -99,9 +83,15 @@ exports.update = (req, res) => {
             message: err.message || "Data to update cannot be empty"
         });
     }
-
-    const id = req.params.id;
-    Userdb.findByIdAndUpdate(id, req.body)
+    const id = req.body.userID;
+    const user = new Userdb({
+        name: req.body.userName,
+        gender: req.body.genderRadio,
+        phone: req.body.phone,
+        email: req.body.email,
+        _id: id
+    })
+    Userdb.findByIdAndUpdate(id, user)
         .then(data => {
             if (!data) {
                 res.status(404).render('error', {
@@ -110,8 +100,8 @@ exports.update = (req, res) => {
             }
             else {
                 console.log(data);
-                res.send(data);
-                //res.redirect('/');
+                //res.send(data);   
+                res.redirect('/admin/users');
             }
         })
         .catch(err => {
@@ -121,10 +111,15 @@ exports.update = (req, res) => {
         });
 }
 
-//delete user with specified user id
+//delete user with specified user id 
 exports.delete = (req, res) => {
-    const id = req.params.id;
-
+    if (!req.body) {
+        return res.status(500).render('error', {
+            message: err.message || "User id to delete cannot be empty"
+        });
+    }
+    const id = req.body.userID;
+    console.log("ID for delete: " + id);
     Userdb.findByIdAndDelete(id)
         .then(data => {
             if (!data) {
@@ -133,8 +128,8 @@ exports.delete = (req, res) => {
                 });
             }
             else {
-                console.log(data);
-                res.redirect('/');
+                console.log("Delete succes: " + data);
+                res.redirect('/admin/users');
             }
         })
         .catch(err => {
@@ -144,16 +139,57 @@ exports.delete = (req, res) => {
         });
 };
 
-exports.login = (body) => {
+//searc user by user name/email 
+exports.search = (req, res) => {
+    if (!req.body) {
+        return res.status(500).render('error', {
+            message: err.message || "Data to search cannot be empty"
+        });
+    }
+    console.log(req.body);
+    const searchValue= req.body.findValue;
+    const regex = new RegExp("^"+searchValue);
+    Userdb.find({ name : { $regex: regex, $options: 'i' }}).sort({name:1}).lean()
+        .then(data => {
+            if (!data) {
+                res.status(404).render('error', {
+                    message: err.message || "Unable to find user"
+                });
+            }
+            else {
+                let empty = false;
+                if(data.length === 0)
+                    empty = true;
+                console.log(data);  
+                res.render('show-users', {
+                    style: 'show-users.css',
+                    title: 'Admin Panel',
+                    navTitle: 'Admin Panel',
+                    loggedIn: true,
+                    users: data,  
+                    showErr: empty
+                });
+            }
+        })
+        .catch(err => {
+            // res.status(500).render('error', {
+            //     message: err.message || "Error searching user in Database"
+            // });
+            res.send(err);
+        });
+}
+
+//login user and admin
+exports.login = (body, val = false) => {
     return new Promise((resolve, reject) => {
         if (!body) {
             reject();
         }
         else {
-            let userData = {};
-            Userdb.findOne({ email: body.email })
+            Userdb.findOne({$and:[{email: body.email}, {isAdmin: val}]})
                 .then(user => {
-                    if (user) {
+                    console.log(user);
+                    if(user !== null) {
                         bcrypt.compare(body.password, user.password)
                             .then((status) => {
                                 if (status)
